Add files via upload
This code defines a node structure (struct node) which has a value (value) and two pointers to other nodes (left and right).
Then, a data type BinaryTree is defined, which is a pointer to such a node.
The following functions are defined:

a) isMinHeap - checks whether a given binary tree is a min-heap. The function takes a binary tree of type BinaryTree and returns a boolean value (true if the tree is a min-heap, false otherwise).
b) biggestMinHeap - calculates the size of the largest min-heap binary subtree in a given binary tree. The function takes a binary tree of type BinaryTree and returns an integer (the size of the largest min-heap binary subtree).
c) biggestMinBinaryHeap - calculates the size of the largest min-heap binary subtree in a given binary search tree. The function takes a binary tree of type BinaryTree and returns an integer (the size of the largest min-heap binary subtree).
d) isBoth - checks whether a given binary tree is a binary search tree. The function takes a binary tree of type BinaryTree and returns a boolean value (true if the tree is a binary search tree, false otherwise).